prisma schema and connnection
