+++ src/app/forms/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <h1>Submit a Project</h1>;
+}
