--- src/utils/db.ts
@@ -1,7 +1,0 @@
-import { neon } from "@neondatabase/serverless";
-
-export async function getData() {
-    const sql = neon(process.env.DATABASE_URL);
-    const data = await sql`SELECT * FROM posts;`;
-    return data;
-}
