--- src/utils/prisma.ts
@@ -1,25 +1,0 @@
-import ws from "ws";
-import { PrismaClient } from "@prisma/client";
-import { PrismaNeon } from "@prisma/adapter-neon";
-import { Pool, neonConfig } from "@neondatabase/serverless";
-
-const prismaClientSingleton = () => {
-  neonConfig.webSocketConstructor = ws;
-  const connectionString = `${process.env.DATABASE_URL}`;
-
-  const pool = new Pool({ connectionString });
-  const adapter = new PrismaNeon(pool);
-  const prisma = new PrismaClient({ adapter });
-
-  return prisma;
-};
-
-declare const globalThis: {
-  prismaGlobal: ReturnType<typeof prismaClientSingleton>;
-} & typeof global;
-
-const prisma = globalThis.prismaGlobal ?? prismaClientSingleton();
-
-export default prisma;
-
-if (process.env.NODE_ENV !== "production") globalThis.prismaGlobal = prisma;
